resolve requested changes
