+++ src/components/GeneralLookupField.tsx
@@ -3,9 +3,11 @@ import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { ResolutionError } from '@unstoppabledomains/resolution/build/resolutionError';
 
 import { DomainStatus, InlineMessage } from '@components';
-import { isValidETHAddress } from '@services/EthService';
-import { isValidENSName } from '@services/EthService/ens';
-import { isValidETHRecipientAddress } from '@services/EthService/validators';
+import {
+  isValidENSName,
+  isValidETHAddress,
+  isValidETHRecipientAddress
+} from '@services/EthService';
 import { getBaseAssetByNetwork, useAssets } from '@services/Store';
 import UnstoppableResolution from '@services/UnstoppableService';
 import { ErrorObject, IReceiverAddress, Network } from '@types';
